connectors, split aspects into files
* Unify the connector interface as `Collector` has
  never been used by itself.
* alerts/silences/states are separate concepts, split
  those into their own files to prepare for adding to
  silences.
