+++ pkg/connectors/silence.go
@@ -1,0 +1,8 @@
+package connectors
+
+import (
+	"context"
+	"time"
+)
+
+type SilencerFunc func(ctx context.Context, duration time.Duration, user string) error
